+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Replace checkout line items amount `code` field with encrypted amount field.
+- Rename checkout lines items `id` field to `product_id`.
+- Refactor checkout line items amount functions.
+
+### Removed
+- Checkout line items data validation.
+- Checkout line items fields mapping.
 
 ## [0.2.0] - 2020-10-04
 ### Added